+++ src/redux/greetings/actions.js
@@ -1,5 +1,5 @@
 export const FETCHDATA = 'GREETINGS/action/FETCHDATA';
 export const fetchdata = (payload) => ({
-    type: FETCHDATA,
-    payload,
-  });
+  type: FETCHDATA,
+  payload,
+});
